+++ src/main/java/uk/co/ogauthority/pwa/model/form/pwaapplications/shared/PermanentDepositsForm.java
@@ -1,5 +1,6 @@
 package uk.co.ogauthority.pwa.model.form.pwaapplications.shared;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 import java.util.Set;
 import uk.co.ogauthority.pwa.model.entity.enums.permanentdeposits.MaterialType;
@@ -22,9 +23,9 @@ public class PermanentDepositsForm extends UploadMultipleFilesWithDescriptionFor
   private String rocksSize;
   private Integer groutBagsSize;
   private String otherMaterialSize;
-  private Integer concreteMattressLength;
-  private Integer concreteMattressWidth;
-  private Integer concreteMattressDepth;
+  private BigDecimal concreteMattressLength;
+  private BigDecimal concreteMattressWidth;
+  private BigDecimal concreteMattressDepth;
 
   private Boolean groutBagsBioDegradable;
   private String bioGroutBagsNotUsedDescription;
@@ -123,27 +124,27 @@ public class PermanentDepositsForm extends UploadMultipleFilesWithDescriptionFor
     this.otherMaterialSize = otherMaterialSize;
   }
 
-  public Integer getConcreteMattressLength() {
+  public BigDecimal getConcreteMattressLength() {
     return concreteMattressLength;
   }
 
-  public void setConcreteMattressLength(Integer concreteMattressLength) {
+  public void setConcreteMattressLength(BigDecimal concreteMattressLength) {
     this.concreteMattressLength = concreteMattressLength;
   }
 
-  public Integer getConcreteMattressWidth() {
+  public BigDecimal getConcreteMattressWidth() {
     return concreteMattressWidth;
   }
 
-  public void setConcreteMattressWidth(Integer concreteMattressWidth) {
+  public void setConcreteMattressWidth(BigDecimal concreteMattressWidth) {
     this.concreteMattressWidth = concreteMattressWidth;
   }
 
-  public Integer getConcreteMattressDepth() {
+  public BigDecimal getConcreteMattressDepth() {
     return concreteMattressDepth;
   }
 
-  public void setConcreteMattressDepth(Integer concreteMattressDepth) {
+  public void setConcreteMattressDepth(BigDecimal concreteMattressDepth) {
     this.concreteMattressDepth = concreteMattressDepth;
   }
